Only validate end date before start date when no other errors
Jamie Jones raised an issue where if the year field was blank the correct error 'You must enter a year' was displayed but also the error 'End date must be after the start date'. After calling the super class, DateValidator, this would find the year empty error and the IllnessEndDateValidator would then still try and validate if end date was before start date as the end date year is set to 2001 when the date fields are extracted from the body even though the year is empty!

Added a check that if there are no validation errors returned from the DateValidator only then validate if the end date is before the start date.

BI-8946